Detect updates
Use a file monitor on /app/.updated to detect when the
flatpak has been updated, and notify the user.

Restart the flatpak

Use the flatpak portal to respawn the latest version
of the app. This requires some gymnastics to make
the running instance quit before we attempt to take
the name.

Use new GApplication api

This does properly what was hacked in before.